added some methods and linemedia model

+++ models/iblock/element.go
@@ -2,7 +2,7 @@ package iblock
 
 type Element struct {
 	Id       int64
-	MinPrice int32
+	MinPrice float64
 	Quantity int
 	Article  string
 	Brand    string
